Update wheatleyDiscord.py
added functionality for new gpt3.5 16k models. !search and !summarize can now pull in a LOT more info to base a response on. The response is still 2000 characters or less to keep under Discords limit. I haven't programmed a way to make it just create a new discord message if it gets close. Maybe soon..